+++ src/main/scala/de/upb/cs/uc4/hyperledger/connections/cases/ConnectionGroup.scala
@@ -7,32 +7,32 @@ import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionGroupTrait
 protected[hyperledger] case class ConnectionGroup(username: String, channel: String, chaincode: String, walletPath: Path, networkDescriptionPath: Path)
   extends ConnectionGroupTrait {
 
-  override def getProposalAddUserToGroup(certificate: String, affiliation: String = AFFILITATION, enrollmentId: String, groupId: String): (String, Array[Byte]) = {
+  override def getProposalAddUserToGroup(certificate: String, affiliation: String = AFFILIATION, enrollmentId: String, groupId: String): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "addUserToGroup", enrollmentId, groupId)
   }
 
-  override def getProposalRemoveUserFromGroup(certificate: String, affiliation: String = AFFILITATION, enrollmentId: String, groupId: String): (String, Array[Byte]) = {
+  override def getProposalRemoveUserFromGroup(certificate: String, affiliation: String = AFFILIATION, enrollmentId: String, groupId: String): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "removeUserFromGroup", enrollmentId, groupId)
   }
 
-  override def getProposalRemoveUserFromAllGroups(certificate: String, affiliation: String = AFFILITATION, enrollmentId: String): (String, Array[Byte]) = {
+  override def getProposalRemoveUserFromAllGroups(certificate: String, affiliation: String = AFFILIATION, enrollmentId: String): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "removeUserFromAllGroups", enrollmentId)
   }
 
-  override def getProposalGetAllGroups(certificate: String, affiliation: String = AFFILITATION): (String, Array[Byte]) = {
+  override def getProposalGetAllGroups(certificate: String, affiliation: String = AFFILIATION): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "getAllGroups")
   }
 
-  override def getProposalGetUsersForGroup(certificate: String, affiliation: String = AFFILITATION, groupId: String): (String, Array[Byte]) = {
+  override def getProposalGetUsersForGroup(certificate: String, affiliation: String = AFFILIATION, groupId: String): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "getUsersForGroup", groupId)
   }
 
-  override def getProposalGetGroupsForUser(certificate: String, affiliation: String = AFFILITATION, enrollmentId: String): (String, Array[Byte]) = {
+  override def getProposalGetGroupsForUser(certificate: String, affiliation: String = AFFILIATION, enrollmentId: String): (String, Array[Byte]) = {
     // TODO: add error handling
     internalGetUnsignedProposal(certificate, affiliation, "getGroupsForUser", enrollmentId)
   }