+++ src/test/scala/de/upb/cs/uc4/hyperledger/tests/ApprovalTests.scala
@@ -21,13 +21,13 @@ class ApprovalTests extends TestBase {
 
   "The ScalaAPI for Approvals" when {
     "invoked for existing contract-transactions " should {
-      "allow for adding new AprovalData " in {
+      "allow for adding new ApprovalData " in {
         chaincodeConnection.approveTransaction("UC4.Certificate", "addCertificate", "000001", "totally valid cert")
       }
-      "allow for adding existing new AprovalData a second time" in {
+      "allow for adding existing new ApprovalData a second time" in {
         chaincodeConnection.approveTransaction("UC4.Certificate", "addCertificate", "000001", "totally valid cert")
       }
-      "allow for adding new AprovalData with wrong number of parameters" in {
+      "allow for adding new ApprovalData with wrong number of parameters" in {
         chaincodeConnection.approveTransaction("UC4.Certificate", "addCertificate", "000001", "totally valid cert", "weird third parameter")
       }
     }